+++ khc/src/khc_state_impl.h
