+++ tests/large_test/khc/post_test.cpp
@@ -1,7 +1,7 @@
 #include "catch.hpp"
 #include <khc.h>
 #include "secure_socket_impl.h"
-#include "khc_impl.h"
+#include "khc_state_impl.h"
 #include "test_callbacks.h"
 #include "picojson.h"
 #include <sstream>
