+++ tests/small_test/khc/minimal_test.cpp
@@ -3,7 +3,7 @@
 #include <string.h>
 #include "catch.hpp"
 #include <khc.h>
-#include "khc_impl.h"
+#include "khc_state_impl.h"
 #include "http_test.h"
 #include "test_callbacks.h"
 #include <sstream>
